move SharedOmniSharpHostFixture to a dedicated file
